Toggle continuous neutral diffusion in interior only
The option to limit neutral diffusion to below the boundary layer
is now implemented by checking to see if the neutral surface
position is within the boundary layer. If so, then put the position
of the neutral position at the nondimensional position and layer of
the bottom boundary layer. This effectively collapses all 'neutral'
surfaces so that layers are only constructed in the ocean interior.