+(*)Add the new parameter VERT_FRICTION_2018_ANSWERS
  Added the new runtime parameter VERT_FRICTION_2018_ANSWERS that avoids the use
of the hard-coded maximum viscous mixing length per timestep in the vertical
viscosity code, and added h_neglect in the denominators of several terms in the
viscosity code.  All answers in the MOM6-examples test cases are bitwise
identical, but the answers will change if ANGSTROM is set to 0, and there is a
new entry in the MOM_parameter_doc files.