* Fixed sign error in thickness diffuse work
This patch fixes a bug in the sign of the tendency of the work by
thickness diffusion in the top layer along the u-points.

This resolves a variance in model runs after a 90-degree rotation, and
results are now consistent with the work_v calculation.

This patch will change answers for any runs using work-based thickness
diffusion.