(*)+Added run-time parameter SURFACE_2018_ANSWERS
  Added a new run-time parameter SURFACE_2018_ANSWERS to improve the calculation
of the surface properties in non-Boussinesq calculations or in the limit where
the total water column is exceptionally thin.  Also corrected a bug in which the
value of src_state%salt_deficit is too large by a factor of 1000000.  This array
does not appear to be used, so this bug might not impact any solutions.  By
default all answers are bitwise identical, but these changes do add and reorder
entries in the MOM_parameter_doc files.