Segment index added for several OBC segment flags
In setup_[uv]_point_obc, a number of flags for each segment were
assigned without index (OBC%segment%flag), which was causing all
segments to be assigned the same value.

This could problems with, say, `[uv]_value_needed` flags, which have
different values depending on if they are E-W or N-S oriented.

This patch adds the index to the assignments.