Started refactoring determine_temperature
  Fixed dimensionally inconsistent expressions in the determine_temperature
routine of midas_vertmap. Thankfully this code had been hard-coded not to be
used, so no answers are changed.  This routine should be carefully reevaluated
and refactored further, as it uses many hard-coded dimensional constants and is
probably not ideal for what it does.  For now, though, all answers are bitwise
identical.