(*)Average Stokes drifts over accumulated depths
  Average Stokes drifts over actual depth of the ocean if it is less than the
averaging depth.  This is not the case in any of the current test case, so
answers do not change.  Also removed two other notes in comments and added
proper dimensional rescaling of one variable in an unused section of the code.
All answers in the test cases are bitwise identical.