+Add REFERENCE_HEIGHT parameter and G%Z_ref
  Added a reference mean sealevel value, G%Z_ref, to the ocean_grid_type,
including the addition of the runtime parameter REFERENCE_HEIGHT and the removal
of the recently added runtime parameter MEAN_SEALEV and the mean_SL entry in the
barotropic_CS.  The mean sea level used in two optionally linearized terms in
the barotropic solver is now taken from G%Z_ref.  All answers are bitwise
identical.