OBC: dz_src rescaling fixed in N-S segments
When reading an OBC segment field from a file which is not V or DVDX,
the field is remapped from the input file levels, set by dz_src, to the
model levels, set by h, via remapping_core_h.

When remapping on east-west segments, dz_src is rescaled to match the
total width of h, saved in variable scl_fac.  But this factor was not
being applied on north-south segments.  This was causing variations in
rotated OBC simulations.

This patch resovles the asymmetry by applying the rescaling along the
N-S segments.