OBC: fix local_open_BC of PPM_reconstruction_y
In PPM_reconstruction_y, the local_open_BC flag was set to the value of
OBC%open_u_BCs_exist_globally, which is for the zonal velocity and is
also used in the PPM_reconstruction_x function.

This discrepancy was causing minor deviations in a dumbbell OBC test and
its 90-degree rotation.

This patch replaces the flag with OBC%open_v_BCs_exist_globally, for the
meridional flow.