Source tree variable fix
This fixes a bug in the SOURCE variable which was used to track files
during development.  It is generally unused for test builds.

This was reworked to include support for FMS and the target repository.
`SOURCE` is now a Makefile function which support MOM6 and FMS trees.
We also extended it to catch other extensions: .inc .c .h

These variables are used to link the top Makefile to the mkmf-generated
makefiles, in order to trigger a rebuild via mkmf-makefile changes.

I would not expect these `SOURCE` dependents to persist if we move to
the autoconf build system.