Modify remapping of vertically extensive diagnostics
Vertically extensive diagnostics like advective transports, diffusive
fluxes, etc. were being remapped onto a diagnostic grid prior to the
physical process that chagne the tracer and/or layer thicknesses.
However this means that the effective operator used for each component
of a tracer budget was not the same, and so budgets could never be
closed cell-by-cell in a diagnostic coordinate.

This commit fixes one part of the inconsistency by setting the target
diagnostic grid for all vertically extensive quantities to be one
constructed at the very beginning of the timestep.

However, problems with closing the budget should still be expected
1. Conservation of column integrals cannot be expected because the
   target grid can be smaller than the source grid (layer thicknesses
   at the native grid at the current point in the algorithm). This
   leads to some fluxes being 'thrown away' due to the reintegrate
   algorithm.
2. To truly have a consistent operator for all terms of a budget,
   the source grid for all vertically extensive grids should also
   be the same