Doxygen 1.8.17 documentation fixes
This resolves several new issues raised by the current version of
Doxygen (1.8.17).  Primarily it addresses the stricter enforcement of
grouping syntax, e.g.

!>@{
...
!>@}

The current code is somewhat inconsistent in its use of !!@{ vs !>@{ ,
and moreso in the closing token.  This patch updates these to always
lead with starting Doxygen comment tokens !> .

Exposing these groups also revealed a few undocumented variables, which
have been updated with minimal descriptions for now.

Finally, one subroutine in MOM_horizontal_regridding was causing a
segfault, so it had to be reworked to remove its grouping, so that
individual indices now have descriptions.