+Corrected diagnostics and diagnostic units
  Corrected the units reported for 9 diagnostics, and altered the code so that
the diagnostics N2_u and N2_v are only offered if the can be calculated and the
proper diagnostics are written if these diagnostics are requested (previously
arrays of zeros were always output).  All solutions are bitwise identical, but
some diagostics in output files change and the available_diags files have
altered entries.