Move allocation of diagnostic arrays
Originally, the diagnostic arrays were being allocated on the first
call generating the diagnostic grid. This seems overly clunky because
the size of the grids are known before that. This was also causing
a segfault in updates to the new routine. The allocate statements
for these arrays are now done right after the number of levels is
known