(*)+Fix grid length use in MOM_ice_shelf_dynamics.F90
  Corrected the grid face lengths used to calculate the mass and tracer fluxes.
This will change answers on a non-Cartesian grid, and because of some bug
corrections, it will change answers if the x- and y- grid spacings are not the
same.  Added comments with the correct units for several variables.  Rearranged
the order of indices in Phi for more efficient memory access.  Applied the MOM6
convention to the case of indicies to indicate grid staggering.  Some variable
names were changed for brevity.  All answers are bitwise identical in the
MOM6-examples test cases, but there are substantial changes to the
MOM_ice_shelf_dynamics.F90 code and it should be noted that there are no active
tests of the ice shelf dynamics code.