(*)Fixed spatial index errors in ice_shelf_advect
  Corrected spatial indexing errors associated with open boundary condition
fluxes in the four ice_shelf_advect routines.  This could change answers in
some cases, but these errors seem likely to lead to segmentation faults in such
cases, so it is entirely possible that this code has never been exercised.
Also revised the directional nomenclature in some comments.  All answers are
bitwise identical in the MOM6-examples test cases, but it should be noted that
there are no active tests of the ice shelf dynamics code.