(*+)Corrected bugs in 3-eqn ice shelf skin salinity
  Corrected several bugs in the 3-equation ice shelf skin salinity calculation,
including renaming variables for greater clarity and using forms for the
solutions to a quadratic equation that are accurate without amplifying roundoff
errors.  In addition, a new runtime parameter, SHELF_3EQ_GAMMA_S, is read and
logged when SHELF_3EQ_GAMMA is true.  This will change answers and the
parameter_doc files with when a thermodynamically active ice shelf is used and
SHELF_THREE_EQN is true, but all answers in the MOM6-examples test cases are
bitwise identical.