(*)Do not use Hml>0 as an ice shelf melt filter
  Avoid using the criterion that Hml>0 as a filter of when ice shelf melt can
occur.  I am not aware of a good justification for this filter, and it seems
to be an historical artefact.  Removing it causes melting to start one time
step earlier at the start of a run, giving apparently better answers.  This
changes the answers in cases with a thermodynamically active ice shelf,
including an ISOMIP test case, but all answers in the MOM6-examples test suite
are bitwise identical.