(*)Removed rescaling from mean_melt_flux to vprec
  Removed inappropriate density ratio rescaling from the conversion of
mean_melt_flux to fluxes%vprec when CONST_SEA_LEVEL is true.  Both are cast as
mass fluxes, not thicknesses fluxes, so this ratio is not needed.  This will
change answers in some regional cases with thermodynamically active ice shelves,
but all answers in the MOM6-examples test suite are bitwise identical.