+Ice shelf code cleanup
  Added a new run-time parameter to specify how much water has to be under an
ice shelf for it to float; this is only logged when the CONST_SEA_LEVEL option
is true.  The description of another parameter is corrected, which changes the
MOM_parameter_doc files with ice shelf thermodynamics.  In addition, merged the
mass_shelf into the same loop as h_shelf in change_thickness_using_melt and
reduced the loop extents for setting forces%p_surf.  Use column masses instead
of thicknesses for thresholds in MOM_ice_shelf.  Also combined CS%Rho0 and
CS%density_ocean_avg as CS%Rho_ocn in ice_shelf_CS.  All answers are bitwise
identical, but there are changes in output files in some cases.