(*)Use state%taux_shelf to set state%ustar_shelf
  If possible, use state%taux_shelf and %tauy_shelf to set state%ustar_shelf,
or if these are not available use the (now appropriately staggered) state%u
and state%v to set ustar%shelf.  In addition, ustar%shelf is set in all cases
with a thermodynamically interactive ice shelf, and not just those that use
the 3-equation expressions for the skin salinity.  In addition, when
CONST_SEA_LEVEL is true, the balancing flux occurs over all open ocean area,
although the previous mode of using a hard-coded region is still there in
commented out code.  These code changes alter answers in all cases with a
thermodynamically interactive ice shelf, but the solutions in the
MOM6-examples test cases are bitwise identical.