+Added dom interface to calculate_density
  Removed the interfaces using G%HI to calculate_density and related routines
and added a new variant with a new optional argument 'dom' specifying a two-
element integer array with the start and end values of the domain to compute
on for 1-d arrays starting at 1.  If this array is not present in this new
variant, calculations are done over the entire output array extent.  All calls
to the interfaces from before the rescale_pressure pull request will still work.
Also added the new function EOS_domain that sets this two-element domain array
from a horiz_index_type.  This new interface is used throughout the code where
the old, removed form was in use.  All answers are bitwise identical.