+Add p_surf to pressures used to calculate density
  Optionally add surface pressure from the ice to equation of state calls.  This
is enabled with the new runtime parameter USE_PSURF_IN_EOS, which is false by
default, preserving the old answers.  These pressures should have been added
all along for physical consistency, so the default should be changed to true and
the option of doing things the previous way should eventually be obsoleted. To
implement this there is a new pointer, tv%p_surf, in the thermo_var_ptrs type,
an EOS_type argument to KPP_compute_BLD was replaced with a thermo_var_ptrs
type, and a new optional argument to neutral_diffusion_calc_coeffs.  By default
all answers are bitwise identical, but there is a new element in a transparent
public type and some minor changes to public interfaces, and there are is a
new entry in the MOM_parameter_doc.all files.