Use deferred sums in MOM_sum_output
  Defer the global sums of fresh_water_in_EFP, etc., from accumulate_net_fluxes
to the point where these accumulated fluxes are used in write_energy, and do the
global sums for up to 5 variables simultaneously.  The reduction in the number
of (blocking) global sums should accelerate the code execution, and because
these calculations use reproducing sums, answers are unchanged by altering the
order of the sums.  All answers are bitwise identical.