+Rescaled visc%MLD and Hml in diabatic to [Z]
  Rescaled the units to [Z] of the visc%MLD element of the vertvisc_type and the
Hml argument to diabatic, which in turn is an argument to bulkmixedlayer,
calculate_CVMix_conv, call_tracer_column_fns, MOM_generic_tracer_column_physics
and mixedlayer_restrat and stored in the MOM_control_struct as CS%Hml.  In
addition when visc%MLD is read from a restart file, it is rescaled if the
scaling has changed between run segments.  All answers are bitwise identical.