Reduced the halo size in an update for eta
  Reduced the halo size for an update to eta to 1.  There had been a note in the
code suggesting that this should be possible for the past 3 years, but that for
unknown reasons this changed the answers with the circle_OBCs test case.  At
some point whatever problems with the OBCs that caused this unexpected
dependency on overly large halos has been corrected, and this halo update can
now use its expected size of 1.  All answers are bitwise identical.