Initialized dMLD_min and dMLD_max in ePBL_column
  Initialized dMLD_min and dMLD_max in ePBL_column, and corrected a comment in
response to helpful reviews from Brandon Reichl and Andrew Shao.  Because these
two arrays are not used until after the 3rd iteration, this may not matter to
the solution, although it should help make the code clearer and avoids unused
variables.  All answers in the MOM6-examples test cases are bitwise identical.