Rotate: fixed-time ALE sponge bugfixes
The sp_val and sp_val_in fields were only allocated for fixed-time
sponges (fixed_sponge=.true.) but were being referenced regardless of
this flag in a few cases.

On x86 architectures this did not create any errors, but was detected on
an Arm64 CPU.

This patch correctly moves all code associated with these variables into
a flag-enabled block.