+Add just_read_params arg to entrain_diffusive_init
  Added the new optional argument just_read_params to entrain_diffusive_init,
which if present and true causes this routine to read the parameters used by the
entrain_diffusive module without logging them.  Also set this new parameter
based on whether ALE remapping is being used.  This prevents the diagnostics and
parameters from this module from being offered or logged when they are not
actually available, and it will prevent the parameter CORRECT_DENSITY from being
logged in ALE-configured runs.  All answers are bitwise identical, but there are
changes in some MOM_parameter_doc and available_diags files.