Add option to apply linear decay at the base of hbl
This patch adds the option to apply a linear decay of
the fluxes at the base of hbl. This had been already
implemented but since it breaks the unit tests, which
were designed to work without this option, adding this
option will avoid breaking the tests.

Also adding minor improvements in the bulk method.