(*)Improve ePBL when EPBL_ORIGINAL_PE_CALC=False
  Modified ePBL_column to calculate Te and Se when EPBL_ORIGINAL_PE_CALC is
False and there are no temperature change diagnostics being requested, and to
return dPEc_dKd from one of the calls to find_PE_chg.  This avoids the use of
uninitialized values in EPBL_ORIGINAL_PE_CALC is false, and some solutions
appear to be similar, but others give large qualitative changes, so there are
probably still problems with EPBL_ORIGINAL_PE_CALC = False.  When this was
originally developed, it was verified to be mathematically equivalent, but in
the years that this code was not tested, problems have crept it.  All answers
in the MOM6-examples test cases are bitwise identical.