+(*)Set the default of DEFAULT_2018_ANSWERS to False
  Changed the default of DEFAULT_2018_ANSWERS to False.  This in turn changes
the default values of 13 other pamarmeters, and could change answers in some
test cases where DEFAULT_2018_ANSWERS or the dependent parameters are not set
explicitly.  By design, the answers in MOM6-examples test suite are bitwise
identical, but they could change in other cases and there are changes to the
MOM_parameter_doc files.