+(*)Changed the default for MEKE_ALPHA_RHINES
  Changed the default value of MEKE_ALPHA_RHINES and MEKE_ALPHA_EADY from 0.05
to 0.0, following the pattern for the defaults for other MEKE parameters.  All
answers in the MOM6-examples test suite are bitwise identical, but this will
change answers in some test cases that use MEKE for diffusivities or viscosities
where MEKE_ALPHA_RHINES or MEKE_ALPHA_EADY are not set explicitly.  There are
changes to the MOM_parameter_doc files.