+(*)Changed WINDSTRESS_STAGGER to WIND_STAGGER
  Changed the parameter to specify the staggering of wind variables in input
files to drive ocean only runs from WINDSTRESS_STAGGER to WIND_STAGGER, with
a default of "C".  This is now consistent with the equivalent parameter name and
default used in the coupled runs.  WINDSTRESS_STAGGER has been obsoleted, and
a new optional warning_val argument has been added to obsolete_char, patterned
after the warning arguments to the obsolete parameter routines for other types.
All answers in the MOM6-examples test suite are bitwise identical, but it is
possible for ocean-only cases reading in A-grid wind stresses to give a fatal
error if their parameters are not updated, and there are changes to the
MOM_parameter_doc files.