+(*)Set the default for KD to 0 m2 s-1
  Set the default value for KD to 0 m2 s-1, and the default value for
KD_KAPPA_SHEAR_0 to the greater of KD and 1e-7 m2 s-1.  Because Kd did not
previously have a default, this can not change answers, but the new default for
KD_KAPPA_SHEAR_0 (which had previously been Kd or 1e-7 m2 s-1 if KD is not set)
could change answers is KD_KAPPA_SHEAR_0 is not set explicitly and KD is set
explicitly to a value that is less than 1e-7 m2 s-1.  All answers in the
MOM-examples test suite are bitwise identical, but there are changes to the
MOM_parameter_doc files.