+(*)Changed default for GRID_ROTATION_ANGLE_BUGS
  Changed the default values of GRID_ROTATION_ANGLE_BUGS and
USE_TRIPOLAR_GEOLONB_BUG to false.  All answers in the MOM6-examples test suite
are bitwise identical, but this will change answers in test cases with the SIS2
sea ice model if these parameters are not set explicitly.  There are changes to
the MOM_parameter_doc and SIS_parameter_doc files.