*Corrected two bugs impacting .testing test cases
  Corrected the velocities that are used in the call to set_viscous_ML with
SPLIT=False and USE_RK2=True.  This call had used the partially updated
predictor velocities, which included other accelerations but not viscosity and
which was missing a halo update, so it would not reproduce between symmetric and
non-symmetric modes or across PE counts.  This combination of parameters is not
used in the MOM6-examples test suite, but it does change answers in some of the
test cases in .testing.  Also removed CORRECT_DENSITY=False from tc4/MOM_input,
so that this test will not fail after this parameter is obsoleted.  All answers
and output in the MOM6-examples test suite are bitwise identical, but there are
deliberate changes in the answers in the .testing test cases.