+Do not log parameters from unused modules
  Modified the get_param calls to avoid logging parameters from MOM_kappa_shear
and MOM_regularize_layers (apart from the main switch to enable these modules)
when those modules are not in use.  There are also two clocks and a diagnostic
that are meaningless when REGULARIZE_SURFACE_LAYERS is false that are no longer
set in that case.  Also corrected two comments in advect_x and advect_y.  All
answers are bitwise identical, but there are changes to some of the
MOM_parameter_doc and available_diags files.