*Fix frazil restarts
  Corrected a bug which causes frazil to be zeroed out during restarts.  This
will change answers in long coupled or ice-ocean runs, but not in the short
MOM6-examples test suite.  This bug was introduced to MOM6 on April 30, 2020
with PR#1102. The answers and documentation files in MOM6-examples are
unchanged.  This PR was tested for restarts, but it used the Baltic test case
and happened to use a day where the frazil was zero everywhere in this limited
test case.  Other test cases should be expected to change answers with every
restart, leading to a small net long-term heat gain in the coupled system.