Testing: Removed environment flag checks
Due to unpredictable conflicts with flags, the realization that flag
testing would launch jobs on Slurm environment, and the fact that Slurm
launchers, OpenMPI, and MPICH all appear to pass environment variables
on default, we now just assume that "env=val $(MPIRUN)" will pass an
environment variable to the MPI job in all cases.

This may not be true on all systems, but for now we will assume this
works.