+Add like_default optional argument to log_param
  Added the new optional argument like_default to the log_param and doc_param
routines to help control where the documentation appears.  This new argument is
used for logging EPBL_USTAR_MIN, the diagnosed output value of MAXIMUM_DEPTH
when the input value is negative, and the diagnosed number of columns where
sponges occur with MOM_ALE_sponge.  An '!' was also added to the logging of
EPBL_USTAR_MIN.  All answers are bitwise identical but there are minor changes
in the contents of some MOM_parameter_doc.short files.