+Add the new parameter INTEGRAL_BT_CONTINUITY
  Add the new runtime parameter INTEGRAL_BT_CONTINUITY which enables the use of
the time-integrated barotropic velocity to determine the cumulative transport
since the start of the barotropic stepping.  This new option works in all of the
MOM6-examples test cases with a SPLIT=True and USE_BT_CONT_TYPE=True.  By
default all answers are bitwise identical, but there are changes to the entries
in the MOM_parameter_doc files.