+Reuse find_uhbt with INTEGRAL_BT_CONT
  Revised the dimensions of the entries in the local_BT_Cont types when
INTEGRAL_BT_CONT=True for efficiency and to enable find_uhbt and other routines
to be used regardless of the value of INTEGRAL_BT_CONT.  Some arguments that are
no longer needed have been removed from some subroutines. All answers are
bitwise identical in the MOM6-examples test suite, but changes to the order of
arithmetic leads to changes from the previous version when
INTEGRAL_BT_CONT=True, and the rescaling is controlled via new optional
arguments to internal routines.