(*?)Revised ice_shelf_driver.F90 so it compiles
  Revised ice_shelf_driver.F90 so that it compiles successfully and is similar
to MOM_driver.F90 where possible.  In addition, deleted the files
MOM_surface_forcing.F90 and user_surface_forcing.F90 in the ice_solo_driver
directory, as these are unused and out-of-date versions of the equivalent files
in the solo_driver directory.  Some of the changes are not as streamlines as
they could be if the solo_ice_shelf code used its own version of diag_mediator,
or if the interface to diag_mediator were revised so that some of its
ocean-specific arguments are optional.  There is no test-case exercising this
code, but at least it now compiles, whereas it did not compile before.