OBC: Define global indices when [IJ]s == [IJ]e
Global indices are well-defined when Is_obc != Ie_obc and Js_obc !=
Je_obc, but not when they are equal.

While the value is unused in these cases, the variables are nontheless
unintialized.

Newer Intel compilers (at least v19+) apply strict assignment tests and
will raise runtime warnings during uninitialized assignment.

We resolve this by assigning the West and South values when the indices
are equal.